refactor: :art: Change of structure in the application folders, using the weekly view as the first interaction.
The folder structure has been changed, separating the weekly and daily views. In addition, the search component now takes you to the results of the weekly weather and within the ability to access the daily information in detail, more generic components have been created for both views and can be reused.